Moved _SetURLDialog to its own module
